drop taskcluster/requirements.*
These files cause us to reinstall taskgraph at the start of each decision task, which is both unnecessary and potentially changes the version being used, which can be confusing.

We do currently use these files as part of a couple of taskgraph tests; I've replaced this usage with installing it through Poetry instead, since we already have taskgraph as a dependency there, and the tests run through Poetry anyways.